Refactor xchain-related code into a separate file.
